Rename util to common
