+++ events/eventer.go
@@ -25,10 +25,10 @@ import (
 	"time"
 
 	"github.com/golang/glog"
+	"k8s.io/heapster/common/flags"
 	"k8s.io/heapster/events/manager"
 	"k8s.io/heapster/events/sinks"
 	"k8s.io/heapster/events/sources"
-	"k8s.io/heapster/util/flags"
 	"k8s.io/heapster/version"
 )
