+++ metrics/sources/kubelet/configs.go
@@ -19,7 +19,7 @@ import (
 	"strconv"
 
 	"github.com/golang/glog"
-	kube_config "k8s.io/heapster/util/kubernetes"
+	kube_config "k8s.io/heapster/common/kubernetes"
 	kube_client "k8s.io/kubernetes/pkg/client/unversioned"
 )
